unify translator tests
